% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-bootstrap.r
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap Methods}
\usage{
bootstrap(x, weights, boot = 100, alpha = 0.05, seed = 120511, bca = FALSE)
}
\arguments{
\item{x}{numeric vector.}

\item{weights}{numeric vector of weight}

\item{boot}{how many bootstrap sample is generated?}

\item{alpha}{significant value to construct confidential interval}

\item{seed}{seed value}

\item{bca}{logical. use bias corrected and accelerated (BCa) bootstrap
to construct confidential intervals.}
}
\description{
Bootstrap Methods
}
\examples{
\dontrun{
x <- rnorm(100)
bootstrap(x, boot = 1000)
}

}
