% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optDiscRD.r
\name{clear_optDiscRD}
\alias{clear_optDiscRD}
\title{Clear Global Options for RCTtoolbox}
\usage{
clear_optDiscRD()
}
\description{
Clear Global Options for RCTtoolbox
}
\examples{
\dontrun{
set_optDiscRD(y1 + y2 ~ run, discRD.cutoff = 45)
clear_optDiscRD()
set_optDiscRD()
}

}
