% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.r
\name{data_bwfilter}
\alias{data_bwfilter}
\title{Subset by Bandwidth}
\usage{
data_bwfilter(data, bw, global = FALSE)
}
\arguments{
\item{data}{list created by \code{clean_rd_data}.}

\item{bw}{numeric vector.
If one element in the vector,
we use data satisfying -abs(bw) <= x & x <= abs(bw).
If missing,
we use treated data and control data closest to cutoff.}

\item{global}{logical.
Whether do you want to restrict data by bandwidth?}
}
\description{
Subset by Bandwidth
}
