% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-lm.r
\name{fit_local_lm}
\alias{fit_local_lm}
\title{Re-fit Local Linear Regression}
\usage{
fit_local_lm(object, extend, force = TRUE)
}
\arguments{
\item{object}{object with fit_wls class}

\item{extend}{numeric vector outside input data}

\item{force}{logical. Whether to ignore error about estimation}
}
\description{
Re-fit Local Linear Regression
}
