% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-lm.r
\name{fit_wls}
\alias{fit_wls}
\title{Fit weighted least squares}
\usage{
fit_wls(y, data, global = TRUE, kernel, bw, point, ...)
}
\arguments{
\item{y}{outcome variable}

\item{data}{data which is generated by \code{\link[=cleand_rd_data]{cleand_rd_data()}}}

\item{global}{logical. use all observations?}

\item{kernel}{character of kernel density ("uniform" or "triangular")}

\item{bw}{bandwidth.}

\item{point}{value of the running variable that
the kernel weights weigh the most.}

\item{\dots}{se_type and cholesky arguments}
}
\description{
Fit weighted least squares
}
