% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discRD-gplot.r
\name{gplot}
\alias{gplot}
\alias{gplot.global_lm}
\title{Flexible RD Plot}
\usage{
gplot(object, ...)

\method{gplot}{global_lm}(
  object,
  order,
  treat_label = c("treated", "control"),
  ate_label_size = 5,
  ate_label_digits = 3,
  ate_label_name = "Local ATE",
  outcome_label,
  ylim,
  vjust = 0,
  hjust = 0,
  xlab = "Running variable",
  ylab = "Average",
  patchwork = TRUE,
  ...
)
}
\arguments{
\item{object}{object}

\item{\dots}{arguments of \code{\link[=simplegg]{simplegg()}}}

\item{order}{numeric. which polynomial order model you use}

\item{treat_label}{string vector of treatment and control label}

\item{ate_label_size}{In-plot text size}

\item{ate_label_digits}{decimal places of in-plot ATE result.}

\item{ate_label_name}{label of in-plot ATE result.}

\item{outcome_label}{Outcome label in plot title}

\item{ylim}{numeric vector of limits of y-axis}

\item{vjust}{numeric. Adjust in-plot text vertically}

\item{hjust}{numeric. Adjust in-plot text horizontally}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}
}
\description{
Flexible RD Plot
}
\examples{
running <- sample(1:100, size = 1000, replace = TRUE)
cov1 <- rnorm(1000, sd = 2); cov2 <- rnorm(1000, mean = -1)
y0 <- running + cov1 + cov2 + rnorm(1000, sd = 10)
y1 <- 2 + 1.5 * running + cov1 + cov2 + rnorm(1000, sd = 10)
y <- ifelse(running <= 50, y1, y0)
bin <- ifelse(y > mean(y), 1, 0)
w <- sample(c(1, 0.5), size = 1000, replace = TRUE)
raw <- data.frame(y, bin, running, cov1, cov2, w)

set_optDiscRD(
  y + bin ~ running,
  discRD.cutoff = 50,
  discRD.assign = "smaller"
)

est <- global_lm(data = raw, weights = w)

gplot(
  est,
  order = 1,
  treat_label = c("Treated", "Control"),
  ylim = list("y" = c(0, 110)),
  vjust = c("y" = 20),
  hjust = c("y" = -0.6),
  ylab = "Weighted Average",
  ate_label_size = 4,
  ate_label_name = "RD estimate"
)

}
