% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-plot.r
\name{label_maker}
\alias{label_maker}
\title{Local ATE label}
\usage{
label_maker(object, label, digits = 3)
}
\arguments{
\item{object}{object with "local_lm" or "global_lm" class}

\item{label}{string of label format}

\item{digits}{decimal place}
}
\description{
Local ATE label
}
