% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-lm.r
\name{lm_internal}
\alias{lm_internal}
\title{Least Squares}
\usage{
lm_internal(y, x, w, cholesky = TRUE, hc = "HC0")
}
\arguments{
\item{y}{outcome vector.}

\item{x}{design matrix}

\item{w}{weight matrix}

\item{cholesky}{logical (default is TRUE).
When solving normal equation, use cholesky decomposition.}

\item{hc}{character.
Calculate robust variance-covariance matrix ("HC0" or "HC1")}
}
\description{
Least Squares
}
\examples{
\dontrun{
x <- cbind(x1 = rnorm(100), x2 = rnorm(100))
y <- x[, 1] + 2 * x[, 2] + rnorm(100)
w <- sample(c(1, 0.5), 100, TRUE)
est <- lm_internal(y, x)
est$estimate
}


}
