% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-random.r
\name{loc_rand_ttest}
\alias{loc_rand_ttest}
\title{Mean difference test (t-test) for local random approach}
\usage{
loc_rand_ttest(y, x = "x", d = "d", w, bw, global = FALSE, bootse, bootp, data)
}
\arguments{
\item{y}{string of outcome variable}

\item{x}{string of running variable}

\item{d}{string of treatment variable}

\item{w}{string of weight variable}

\item{bw}{numeric vector of bandwidth.
If specified, use data
whose running variables are within this range will be used.
If missing, use data from treatment and control groups
where the running variable is closest to the cutoff}

\item{global}{logical (default is FALSE).
Whether to use all observations.}

\item{bootse}{numeric.
Generate N averages of bootstrap samples and
use the standard deviation as the standard error of the average.
If missing, standard error of mean is calculated by
sqrt(v / e) where v is unbiased variance of outcome,
and e is effective sample size: e = sum(w) ^ 2 / sum(w ^ 2).
If w is missing, we set w = 1, and obtain e = 1 / n.
standard error of mean difference is obtained by
sqrt((v1 / e1) + (v0 / e0))}

\item{bootp}{numeric.
Perform a permutation test with N re-randomizations.
The p-value is obtained at a rate
where the absolute value of the mean difference
due to rerandomization is greater than
the absolute value of the observed mean difference.
If missing, standard t-test is performed.}

\item{data}{data.frame which you want to use.}
}
\description{
Mean difference test (t-test) for local random approach
}
\examples{
\dontrun{
loc_rand_ttest("y", data = dt2)
loc_rand_ttest("y", global = TRUE, data = dt2)
loc_rand_ttest("y", bootse = 100, data = dt2)
loc_rand_ttest("y", bootp = 100, data = dt2)
loc_rand_ttest("y", bw = c(-5, 5), bootp = 100, data = dt2)
}

}
