% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-random.r
\name{local_random}
\alias{local_random}
\title{Perform Local Randomization Approach}
\usage{
local_random(
  y,
  x,
  cutoff,
  assign,
  w,
  bw,
  global = FALSE,
  bootse,
  bootp,
  data,
  suby
)
}
\arguments{
\item{y}{string vector with outcome variables
If missing, try to find \code{getOption("discRD.outcome")}.}

\item{x}{string of running variable
If missing, try to find \code{getOption("discRD.running")}.}

\item{cutoff}{numeric of cutoff points
If missing, try to find \code{getOption("discRD.cutoff")}.}

\item{assign}{assignment rule of treatment.
If "greater",
treated whose running variable is greater than or equal to cutoff.
If "smaller",
treated whose running variable is less than or equal to cutoff.
If missing, try to find \code{getOption("discRD.assign")}.}

\item{w}{string of weight variable}

\item{bw}{numeric vector of bandwidth.
If specified, use data
whose running variables are within this range will be used.
If missing, use data from treatment and control groups
where the running variable is closest to the cutoff}

\item{global}{logical (default is FALSE).
Whether to use all observations.}

\item{bootse}{numeric.
Generate N averages of bootstrap samples and
use the standard deviation as the standard error of the average.
If missing, standard error of mean is calculated by
sqrt(v / e) where v is unbiased variance of outcome,
and e is effective sample size: e = sum(w) ^ 2 / sum(w ^ 2).
If w is missing, we set w = 1, and obtain e = 1 / n.
standard error of mean difference is obtained by
sqrt((v1 / e1) + (v0 / e0))}

\item{bootp}{numeric.
Perform a permutation test with N re-randomizations.
The p-value is obtained at a rate
where the absolute value of the mean difference
due to rerandomization is greater than
the absolute value of the observed mean difference.
If missing, standard t-test is performed.}

\item{data}{data.frame which you want to use.}

\item{suby}{numeric vector.
Specify position of outcome variables you want to use.}
}
\description{
An analysis of the difference between
the means of the two groups is performed
under the assumption that
the treatment assignments are random near the cutoff (local random).
Statistical inference has a permutation test for small sample sizes
in addition to the standard t-test.
}
\examples{
running <- sample(1:100, size = 1000, replace = TRUE)
cov1 <- rnorm(1000, sd = 2); cov2 <- rnorm(1000, mean = -1)
y0 <- running + cov1 + cov2 + rnorm(1000)
y1 <- 2 + 1.5 * running + cov1 + cov2 + rnorm(1000)
y <- ifelse(running <= 50, y1, y0)
bin <- ifelse(y > mean(y), 1, 0)
raw <- data.frame(y, bin, running, cov1, cov2)

set_optDiscRD(
  y + bin ~ running,
  discRD.cutoff = 50,
  discRD.assign = "smaller"
)

local_random(data = raw)
local_random(data = raw, bootp = 100, bootse = 100, suby = 1)
local_random("y", data = raw, bootp = 100)

}
