% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discRD-local-random.r
\name{local_random_test_int}
\alias{local_random_test_int}
\title{Mean difference test (t-test) for local random approach}
\usage{
local_random_test_int(basemod, data, bootse, bootp)
}
\arguments{
\item{basemod}{baseline formula. \verb{outcome ~ running variable}.}

\item{data}{data.frame which you want to use.}

\item{bootse}{numeric.
Generate N averages of bootstrap samples and
use the standard deviation as the standard error of the average.
If missing, standard error of mean is calculated by
sqrt(v / e) where v is unbiased variance of outcome,
and e is effective sample size: e = sum(w) ^ 2 / sum(w ^ 2).
If w is missing, we set w = 1, and obtain e = 1 / n.
standard error of mean difference is obtained by
sqrt((v1 / e1) + (v0 / e0))}

\item{bootp}{numeric.
Perform a permutation test with N re-randomizations.
The p-value is obtained at a rate
where the absolute value of the mean difference
due to rerandomization is greater than
the absolute value of the observed mean difference.
If missing, standard t-test is performed.}
}
\description{
Mean difference test (t-test) for local random approach
}
\examples{
\dontrun{
running <- sample(1:100, size = 1000, replace = TRUE)
cov1 <- rnorm(1000, sd = 2); cov2 <- rnorm(1000, mean = -1)
y0 <- running + cov1 + cov2 + rnorm(1000, sd = 10)
y1 <- 2 + 1.5 * running + cov1 + cov2 + rnorm(1000, sd = 10)
y <- ifelse(running <= 50, y1, y0)
bin <- ifelse(y > mean(y), 1, 0)
w <- sample(c(1, 0.5), size = 1000, replace = TRUE)
raw <- data.frame(y, bin, running, cov1, cov2, w)

set_optDiscRD(discRD.cutoff = 50, discRD.assign = "smaller")
useit <- clean_rd_data(y ~ running, data = raw)
str(local_random_test_int(y ~ running, useit$data))
str(local_random_test_int(y ~ running, useit$data, bootse = 100))
str(local_random_test_int(y ~ running, useit$data, bootp = 100))
}

}
