% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-func.r
\name{permutation}
\alias{permutation}
\title{Permutation test}
\usage{
permutation(y, d, weights, boot = 100, seed = 120511)
}
\arguments{
\item{y}{numeric vector of outcome}

\item{d}{numeric vector of treatment indicator (contain 1 and 0 only)}

\item{weights}{numeric vector of weights}

\item{boot}{how many re-randomization are generated}

\item{seed}{seed value}
}
\description{
Permutation test
}
\examples{
\dontrun{
y1 <- 2 + rnorm(20, sd = 4)
y0 <- rnorm(20, sd = 4)
d <- sample(c(1, 0), replace = TRUE, size = 40)
y <- ifelse(d == 1, y1, y0)
permutation(y, d, boot = 200)
}

}
