% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discRD-rdplot.r
\name{rdplot}
\alias{rdplot}
\alias{rdplot.list_global_lm}
\alias{rdplot.list_local_lm}
\title{Flexible RD Plot}
\usage{
rdplot(object, ...)

\method{rdplot}{list_global_lm}(
  object,
  usemod,
  treat_label = c("treated", "control"),
  ate_label_size = 5,
  ate_label_digits = 3,
  ate_label_format,
  outcome_label,
  ylim,
  vjust = 0,
  hjust = 0,
  xlab = "Running variable",
  ylab = "Average",
  patchwork = TRUE,
  ncol = NULL,
  nrow = NULL,
  ...
)

\method{rdplot}{list_local_lm}(
  object,
  usemod,
  treat_label = c("treated", "control"),
  ate_label_size = 5,
  ate_label_digits = 3,
  ate_label_format,
  outcome_label,
  ylim,
  vjust = 0,
  hjust = 0,
  xlab = "Running variable",
  ylab = "Average",
  patchwork = TRUE,
  ncol = NULL,
  nrow = NULL,
  force = TRUE,
  ...
)
}
\arguments{
\item{object}{object}

\item{\dots}{arguments of \code{\link[=simplegg]{simplegg()}}}

\item{usemod}{numeric vector. which models do you want to plot?}

\item{treat_label}{string vector of treatment and control label}

\item{ate_label_size}{In-plot text size}

\item{ate_label_digits}{decimal places of in-plot ATE result.}

\item{ate_label_format}{string of label format of in-plot ATE result.
The format produces labels according to a unique grammar.
If you want to embed a numerical value related to the estimation result,
you need to enclose it in {}.
\itemize{
\item \code{{estimate}}: embed coefficient value (Local ATE)
\item \code{{std.error}}: embed standard error of coefficient
\item \code{{statistic}}: embed z-score of coefficient
\item \code{{p.value}}: embed p-value of coefficient
\item \code{{star}}: show *** if p-value <= 0.01;
** if p-value <= 0.05; * if p-value <= 0.1
}}

\item{outcome_label}{Outcome label in plot title}

\item{ylim}{numeric vector of limits of y-axis}

\item{vjust}{numeric. Adjust in-plot text vertically}

\item{hjust}{numeric. Adjust in-plot text horizontally}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{patchwork}{logical.
add a list of plots into one composition by \code{patchwork::wrap_plots}}

\item{ncol}{the dimensions of the grid to create.
argument of \code{patchwork::wrap_plots}}

\item{nrow}{the dimensions of the grid to create.
argument of \code{patchwork::wrap_plots}}

\item{force}{logical. Whether to ignore error about estimation}
}
\description{
Flexible RD Plot
}
\examples{
running <- sample(1:100, size = 1000, replace = TRUE)
cov1 <- rnorm(1000, sd = 2); cov2 <- rnorm(1000, mean = -1)
y0 <- running + cov1 + cov2 + rnorm(1000, sd = 10)
y1 <- 2 + 1.5 * running + cov1 + cov2 + rnorm(1000, sd = 10)
y <- ifelse(running <= 50, y1, y0)
bin <- ifelse(y > mean(y), 1, 0)
w <- sample(c(1, 0.5), size = 1000, replace = TRUE)
raw <- data.frame(y, bin, running, cov1, cov2, w)

set_optDiscRD(
  y + bin ~ running,
  discRD.cutoff = 50,
  discRD.assign = "smaller"
)

est <- global_lm(data = raw, weights = w)

rdplot(
  est,
  usemod = c(1, 2),
  treat_label = c("Treated", "Control"),
  outcome_label = switch(x,
    "1" = "Simulated outcome",
    "2" = "Simulated outcome > 0"
  ),
  ylab = "Weighted Average",
  ate_label_size = 4
)

nonpara <- local_lm(
  submod = 1,
  data = raw,
  kernel = "uniform",
  bw = 5,
  order = 1:3
)

rdplot(
  nonpara,
  usemod = 2,
  outcome_label = switch(x,
    "1" = "Order = 1",
    "2" = "Order = 2",
    "3" = "Order = 3"
  )
)

}
