% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-plot.r
\name{rdplot_internal_cutoff}
\alias{rdplot_internal_cutoff}
\title{Basic RD plot}
\usage{
rdplot_internal_cutoff(
  aggregate,
  predict1,
  predict0,
  cutoff,
  ate_label,
  ate_label_size = 5,
  outcome_label,
  ylim,
  vjust = 0,
  hjust = 0,
  xlab = "Running variable",
  ylab = "Average",
  ...
)
}
\arguments{
\item{aggregate}{observed data aggregated by mass points}

\item{predict1}{predicted data of treated}

\item{predict0}{predicted data of control}

\item{cutoff}{cutoff value}

\item{ate_label}{In-plot text about local ATE estimates}

\item{ate_label_size}{In-plot text size}

\item{outcome_label}{Outcome label in plot title}

\item{ylim}{numeric vector of limits of y-axis}

\item{vjust}{numeric. Adjust in-plot text vertically}

\item{hjust}{numeric. Adjust in-plot text horizontally}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{\dots}{arguments of \code{\link[=simplegg]{simplegg()}}}
}
\description{
Basic RD plot
}
