% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discRD-tabular.r
\name{rdtab.list_global_lm}
\alias{rdtab.list_global_lm}
\title{Output Table for Global Local Polynomials}
\usage{
\method{rdtab}{list_global_lm}(
  object,
  ylab,
  dlab = "treated",
  olab = "Order of polynomial",
  covariate_labs,
  stars = c(`***` = 0.01, `**` = 0.05, `*` = 0.1),
  title = NULL,
  footnote = NULL,
  output = getOption("discRD.table_output"),
  fontsize = getOption("discRD.table_fontsize"),
  digits = 3,
  ...
)
}
\arguments{
\item{object}{object with "global_lm" class}

\item{ylab}{a named string vector of outcome variables}

\item{dlab}{a string of label of treatment variable}

\item{olab}{a string of label of "Order of polynomial"}

\item{covariate_labs}{list.
You can specify how to display information about
whether you are controlling a variable that is not displayed by a factor.
For example,
\code{list("label 1" = "x1")} changes the label of "x1" to "label 1".
Also, when specified as \code{list("label" = c ("x1", "x2"))},
if x1 and x2 are not controlled or
the coefficients are not displayed in all models,
"x1" and "x2" labels are aggregated into one line with the name "label".}

\item{stars}{a named numeric vector to indicate statistical significance.
\code{"***" = 0.01} means show *** if p-value is less than or equal to 0.01.}

\item{title}{a string of title}

\item{footnote}{a string of footnote}

\item{output}{a string of output format.
If missing, find \code{options("discRD.table_output")}.}

\item{fontsize}{numeric of font size.
If missing, find \code{options("discRD.table_fontsize")}.}

\item{digits}{numeric.
the number of digits to keep after the period.}

\item{\dots}{Other arguments to pass to \code{kableExtra::kable_styling}}
}
\description{
Output Table for Global Local Polynomials
}
\examples{
\dontrun{
running <- sample(1:100, size = 1000, replace = TRUE)
cov1 <- rnorm(1000, sd = 2); cov2 <- rnorm(1000, mean = -1)
y0 <- running + cov1 + cov2 + rnorm(1000, sd = 10)
y1 <- 2 + 1.5 * running + cov1 + cov2 + rnorm(1000, sd = 10)
y <- ifelse(running <= 50, y1, y0)
bin <- ifelse(y > mean(y), 1, 0)
w <- sample(c(1, 0.5), size = 1000, replace = TRUE)
raw <- data.frame(y, bin, running, cov1, cov2, w)

set_optDiscRD(
  y + bin ~ running,
  xmod = ~ cov1 + cov2,
  discRD.cutoff = 50,
  discRD.assign = "smaller"
)

library(magrittr)

global_lm(data = raw) \%>\%
  tabular(
    title = "Estimate Local ATE by Global Polynomial Fitting",
    ylab = c("y" = "Simulated Outcome", "bin" = "Simulated Outcome > 0"),
    dlab = "Treatment",
    covariate_labs = list("Covariates" = c("cov1", "cov2")),
    footnote = "***: p < 0.01, **: p < 0.05, *: p < 0.1"
  )
}

}
