% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discRD-rdtab.r
\name{rdtab.list_local_random}
\alias{rdtab.list_local_random}
\title{Output Table for Local Random Approach}
\usage{
\method{rdtab}{list_local_random}(
  object,
  ylab,
  outcome_lab = "Outcomes",
  meanlab = "Mean",
  selab = "S.E.",
  nlab = "N",
  meandiff_lab = "Mean difference",
  plab = "P-value",
  treat_header = "Treated",
  control_header = "Control",
  title = NULL,
  footnote = NULL,
  output = getOption("discRD.table_output"),
  size = getOption("discRD.table_fontsize"),
  digits = 3,
  ...
)
}
\arguments{
\item{object}{object with "local_random_test" class}

\item{ylab}{string vector with outcome labels.
Specify c("original label" = "new label", ...)}

\item{outcome_lab}{string of label of outcome column.}

\item{meanlab}{string of label of columns showing average.}

\item{selab}{string of label of columns showing standard error.}

\item{nlab}{string of label of columns showing number of observations.}

\item{meandiff_lab}{string of label of columns showing
mean difference.}

\item{plab}{string of label of columns showing p-value of
t-test (or permutation test).}

\item{treat_header}{string of label of treatment}

\item{control_header}{string of label of control}

\item{title}{string of table title}

\item{footnote}{string of footnote}

\item{output}{string of output format
(default is getOption("discRD.table_output"))}

\item{size}{numeric of font size
(default is getOption("discRD.table_fontsize"))}

\item{digits}{numeric of number of decimal places to display
(Default is 3).}

\item{\dots}{Other arguments to pass to \code{kableExtra::kable_styling}}
}
\description{
Output Table for Local Random Approach
}
\examples{
running <- sample(1:100, size = 1000, replace = TRUE)
cov1 <- rnorm(1000, sd = 2); cov2 <- rnorm(1000, mean = -1)
y0 <- running + cov1 + cov2 + rnorm(1000)
y1 <- 2 + 1.5 * running + cov1 + cov2 + rnorm(1000)
y <- ifelse(running <= 50, y1, y0)
bin <- ifelse(y > mean(y), 1, 0)
raw <- data.frame(y, bin, running, cov1, cov2)

set_optDiscRD(
  y + bin ~ running,
  discRD.cutoff = 50,
  discRD.assign = "smaller"
)

library(modelsummary)
library(magrittr)

ylab <- c("y" = "Simulated Outcome", "bin" = "Binary")
footnote <- paste0(
  "Observations with running variables of -45 to 55 were used.",
  "The treatment variable is determined to take 1",
  "if the running variable is 50 or less."
)

local_random_test(data = raw, bw = c(-5, 5)) \%>\%
  rdtab(ylab, digits = 2, footnote = footnote)

}
