% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discRD-tabular.r
\name{rdtab}
\alias{rdtab}
\title{Output Table of Empirical Analysis for Discrete RD}
\usage{
rdtab(...)
}
\arguments{
\item{\dots}{some arguments.
See \code{\link[=tabular.local_random_test]{tabular.local_random_test()}} for arguments of "local_random" class.
See \code{\link[=tabular.global_lm]{tabular.global_lm()}}
for arguments of "discRD_global_lm" class.}
}
\description{
Output Table of Empirical Analysis for Discrete RD
}
