% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-recover.r
\name{recover_data}
\alias{recover_data}
\alias{recover_data.list_global_lm}
\title{Recover Data from estimation object}
\usage{
recover_data(object, ...)

\method{recover_data}{list_global_lm}(object, modnum)
}
\arguments{
\item{object}{object with global_lm class}

\item{\dots}{other arguments}

\item{modnum}{numeric. which model you use.}
}
\description{
Recover Data from estimation object
}
