% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optDiscRD.r
\name{set_optDiscRD}
\alias{set_optDiscRD}
\title{Manipulation Global Options for {discRD}}
\usage{
set_optDiscRD(basicmod, xmod, ...)
}
\arguments{
\item{basicmod}{two-sided formula with
outcome variables on lhs and running variable on rhs}

\item{xmod}{(list of) one-sided formulas with
covariates used in regression analysis on rhs.}

\item{\dots}{Specify \code{option_name = option}}
}
\description{
Manipulation Global Options for {discRD}
}
\details{
If empty arguments, return current options.
}
\examples{
\dontrun{
set_optDiscRD(y1 + y2 ~ run, discRD.cutoff = 45)
clear_optDiscRD()
set_optDiscRD()
}

}
