% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{shape_data}
\alias{shape_data}
\title{Data Shape}
\usage{
shape_data(y, x, cutoff, assign, data, ...)
}
\arguments{
\item{y}{string vector with outcome variables}

\item{x}{string with running variable}

\item{cutoff}{numeric of cutoff points}

\item{assign}{assignment rule of treatment.
If "greater",
treated whose running variable is greater than or equal to cutoff.
If "smaller",
treated whose running variable is less than or equal to cutoff.}

\item{data}{original data.frame}

\item{\dots}{other string vector with variables to keep}
}
\description{
Data Shape
}
\examples{
\dontrun{
running <- sample(1:100, size = 1000, replace = TRUE)
cov1 <- rnorm(1000, sd = 2); cov2 <- rnorm(1000, mean = -1)
y0 <- running + cov1 + cov2 + rnorm(1000)
y1 <- 2 + 1.5 * running + cov1 + cov2 + rnorm(1000)
y <- ifelse(running <= 50, y1, y0)
raw <- data.frame(y, running, cov1, cov2)

with(raw, plot(running, y, pch = ifelse(running <= 50, 16, 17)))

dt <- shape_data("y", "running", 50, "smaller", raw)
with(dt, plot(x, y, pch = ifelse(d == 1, 16, 17)))

set_optDiscRD(
  y ~ running,
  discRD.cutoff = 50,
  discRD.assign = "smaller"
)
dt2 <- shape_data(data = raw, cov = c("cov1", "cov2"))
with(dt2, plot(x, y, pch = ifelse(d == 1, 16, 17)))
}


}
