% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplegg.r
\name{simplegg}
\alias{simplegg}
\title{Template for {ggplot2}}
\usage{
simplegg(
  title_size = 13,
  axis_title_size = 13,
  axis_text_size = 12,
  legend_title_size = 12,
  legend_text_size = 12,
  caption_size = 11,
  legend_key_size = 1,
  font_family = NULL,
  flip = FALSE
)
}
\arguments{
\item{title_size}{numeric. Specify font size of plot title.}

\item{axis_title_size}{numeric. Specify font size of axis title.}

\item{axis_text_size}{numeric. Specify font size of axis texts.}

\item{legend_title_size}{numeric. Specify font size of legend title.}

\item{legend_text_size}{numeric. Specify font size of legend texts.}

\item{caption_size}{numeric. Specify font size of caption.}

\item{legend_key_size}{numeric. Specify key size of legends}

\item{font_family}{character. Specify font family. Default is NULL.}

\item{flip}{logical. Specify whether to swap the x-axis and y-axis.}
}
\description{
Template for {ggplot2}
}
