% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-method.r
\name{summary.local_random_test}
\alias{summary.local_random_test}
\title{Summary function}
\usage{
\method{summary}{local_random_test}(object, ...)
}
\arguments{
\item{object}{object with "local_random_test" class}

\item{\dots}{additional arguments affecting the summary produced.}
}
\description{
Summary function
}
