% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discRD-tab.r
\name{tabular}
\alias{tabular}
\title{Output Table of Empirical Analysis for Discrete RD}
\usage{
tabular(...)
}
\arguments{
\item{\dots}{some arguments.
See \code{\link[=tabular.local_random_test]{tabular.local_random_test()}} for arguments of "local_random" class.
See \code{\link[=tabular.discRD_global_lm]{tabular.discRD_global_lm()}}
for arguments of "discRD_global_lm" class.}
}
\description{
Output Table of Empirical Analysis for Discrete RD
}
