% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-methods.r
\name{tidy.local_ate_global_lm}
\alias{tidy.local_ate_global_lm}
\title{Methods of tidy and glance}
\usage{
\method{tidy}{local_ate_global_lm}(x, ...)
}
\arguments{
\item{x}{object with local_ate_global_lm class}

\item{\dots}{other arguments}
}
\description{
Methods of tidy and glance
}
