% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-clean.r
\name{clean_rd_data}
\alias{clean_rd_data}
\title{Data Shape}
\usage{
clean_rd_data(
  basemod,
  covmod,
  data,
  subset,
  weights,
  cluster,
  order = 1,
  cutoff,
  assign,
  bw,
  global = TRUE
)
}
\arguments{
\item{basemod}{baseline formula. \verb{outcome ~ running variable}.}

\item{covmod}{one-sided formula with covariates on rhs.}

\item{data}{data.frame}

\item{subset}{subset condition.}

\item{weights}{weight variable.}

\item{cluster}{cluster variable.}

\item{order}{order of polynomial.}

\item{cutoff}{numeric of cutoff point.
If missing, search \code{option("discRD.cutoff")}}

\item{assign}{assignment rule of treatment.
If "greater",
treated whose running variable is greater than or equal to cutoff.
If "smaller",
treated whose running variable is less than or equal to cutoff.
If missing, search \code{option("discRD.assign")}}

\item{bw}{numeric vector of bandwidth.
If specified, use data
whose running variables are within this range will be used.
If missing, use data from treatment and control groups
where the running variable is closest to the cutoff}

\item{global}{logical (default is FALSE).
Whether to use all observations.}
}
\description{
Data Shape
}
\examples{
\dontrun{
running <- sample(1:100, size = 1000, replace = TRUE)
cov1 <- rnorm(1000, sd = 2); cov2 <- rnorm(1000, mean = -1)
y0 <- running + cov1 + cov2 + rnorm(1000, sd = 10)
y1 <- 2 + 1.5 * running + cov1 + cov2 + rnorm(1000, sd = 10)
y <- ifelse(running <= 50, y1, y0)
bin <- ifelse(y > mean(y), 1, 0)
w <- sample(c(1, 0.5), size = 1000, replace = TRUE)
raw <- data.frame(y, bin, running, cov1, cov2, w)

set_optDiscRD(discRD.cutoff = 50, discRD.assign = "smaller")
a <- clean_rd_data(y ~ running + cov1, data = raw, weights = w, order = 2)
str(a)
}


}
