% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{package}
\name{discreteRD}
\alias{discreteRD}
\alias{discreteRD-package}
\title{Useful Wrapper Functions for Discrete Regression Discontinuity Design}
\description{
What the package does (one paragraph).
}
\section{Package Options}{


discreteRD provides the following \code{\link[=options]{options()}}
to facilitate consistent analysis:
\itemize{
\item \code{discRD.basemod}: a list of formula with \code{outcome ~ running}.
\item \code{discRD.cutoff}: a numeric value of cutoff point.
\item \code{discRD.assign}: a string of assignment rule.
If "greater", treatment is determined by 1*(running >= cutoff).
If "smaller", treatment is determined by 1*(running <= cutoff).
\item \code{discRD.covmod}: A list of one-sided formulas with covariates
on the right-hand side of the equation
\item \code{discRD.unique.covariate}: String vector with unique covariates.
\item \code{discRD.plot_family}: The name of the font family
used to output the figure with {ggplot2}.
\item \code{discRD.table_output}: String of output format with {modelsummary}.
Default is "kableExtra".
\item \code{discRD.table_fontsize}: The font size in the table
output by {kableExtra} and {flextable} used through {modelsummary}.
}
}

\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
