% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discRD-global-lm.r
\name{global_lm}
\alias{global_lm}
\title{Global Linear Least Squares to Estimate Local ATE}
\usage{
global_lm(
  basemod,
  covmod,
  data,
  subset,
  weights,
  cluster,
  submod,
  onlydmod = TRUE,
  order = c(1, 2),
  cutoff,
  assign,
  se = "HC0",
  cholesky = TRUE
)
}
\arguments{
\item{basemod}{baseline formula. \verb{outcome ~ running variable}.}

\item{covmod}{(list of) one-sided formula with
covariates on rhs.
If NULL, covariates are not controlled.
If missing, find \code{options("discRD.xmod")}}

\item{data}{data.frame which you want to use.}

\item{subset}{subset condition.}

\item{weights}{weight variable.}

\item{cluster}{cluster variable.}

\item{submod}{numeric vector.
Which baseline model you want to use.}

\item{onlydmod}{logical (default is TRUE).
Whether to estimate a model without covariates.}

\item{order}{numeric vector of global polynomial orders.}

\item{cutoff}{numeric of cutoff points
If missing, try to find \code{getOption("discRD.cutoff")}.}

\item{assign}{assignment rule of treatment.
If "greater",
treated whose running variable is greater than or equal to cutoff.
If "smaller",
treated whose running variable is less than or equal to cutoff.
If missing, try to find \code{getOption("discRD.assign")}.}

\item{se}{character.
How to calculate robust variance-covariance matrix
("HC0", "HC1", "HC2", "HCj", "HC3", and "HC4")
If "standard", calculate conventional variance-covariance matrix}

\item{cholesky}{logical (default is TRUE).
When solving normal equation, use cholesky decomposition.}
}
\description{
Estimate the local ATE
at the treatment assignment threshold
by the global least squares method.
}
\examples{
running <- sample(1:100, size = 1000, replace = TRUE)
cov1 <- rnorm(1000, sd = 2); cov2 <- rnorm(1000, mean = -1)
y0 <- running + cov1 + cov2 + rnorm(1000, sd = 10)
y1 <- 2 + 1.5 * running + cov1 + cov2 + rnorm(1000, sd = 10)
y <- ifelse(running <= 50, y1, y0)
bin <- ifelse(y > mean(y), 1, 0)
w <- sample(c(1, 0.5), size = 1000, replace = TRUE)
raw <- data.frame(y, bin, running, cov1, cov2, w)

set_optDiscRD(
  y + bin ~ running,
  covmod = list(~cov1, ~ cov1 + cov2),
  discRD.cutoff = 50,
  discRD.assign = "smaller"
)

est <- global_lm(data = raw)
str(global_lm(submod = 1, data = raw))
str(global_lm(submod = 1, order = 3, data = raw))
str(global_lm(data = raw, cutoff = 30))
est2 <- global_lm(data = raw, se = "HC1", weights = w)

}
