% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-lm.r
\name{wls}
\alias{wls}
\title{Weighted Least Squares}
\usage{
wls(y, x, w, se = "HC0", cluster, cholesky = TRUE)
}
\arguments{
\item{y}{outcome vector.}

\item{x}{design matrix}

\item{w}{weight matrix}

\item{se}{character.
How to calculate robust variance-covariance matrix
("HC0", "HC1", "HC2", "HCj", "HC3", and "HC4")
If "standard", calculate conventional variance-covariance matrix}

\item{cluster}{cluster variable.}

\item{cholesky}{logical (default is TRUE).
When solving normal equation, use cholesky decomposition.}
}
\description{
Weighted Least Squares
}
\examples{
\dontrun{
x <- cbind(x1 = rnorm(100), x2 = rnorm(100))
y <- x[, 1] + 2 * x[, 2] + rnorm(100)
w <- sample(c(1, 0.5), 100, TRUE)
est <- lm_internal(y, x)
est$estimate
}


}
